[GenericModel] Set path to "0" if iterator is not set
It's not clear why get_path() is called with a tree_iter that is not
yet defined -- in other words there may be some application logic bug
hidden here -- but this should make the call safe: Return the "0" path
when iterator is not initialized (tree_iter.user_data is null or 0,
depending on gobject-introspection version).

Fixes #38, I believe. I cannot reproduce the problem as is but I can manually
create problematic tree_iters: those are now fine with the new code.

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>